Ejercicio 3 completo

+++ tp2_FloresRodrigo/src/ar/edu/unju/fi/ejercicio3/constantes/Provincia.java
@@ -1,15 +1,16 @@
 package ar.edu.unju.fi.ejercicio3.constantes;
 
 public enum Provincia {
-	JUJUY(811611, 53219), SALTA(1441351, 155488), TUCUMAN(1731820, 22524), CATAMARCA(429562, 102602), LA_RIOJA(383865, 89680), SANTIAGO_DEL_ESTERO(1060906, 136351);
-	
+	JUJUY(811611, 53219), SALTA(1441351, 155488), TUCUMAN(1731820, 22524), CATAMARCA(429562, 102602),
+	LA_RIOJA(383865, 89680), SANTIAGO_DEL_ESTERO(1060906, 136351);
+
 	private int poblacion;
-	private double superficie;
-	
+	private float superficie;
+
 	private Provincia() {
 	}
 
-	private Provincia(int poblacion, double superficie) {
+	private Provincia(int poblacion, float superficie) {
 		this.poblacion = poblacion;
 		this.superficie = superficie;
 	}
@@ -22,15 +23,15 @@ public enum Provincia {
 		this.poblacion = poblacion;
 	}
 
-	public double getSuperficie() {
+	public float getSuperficie() {
 		return superficie;
 	}
 
-	public void setSuperficie(double superficie) {
+	public void setSuperficie(float superficie) {
 		this.superficie = superficie;
 	}
-	
-	public double calcularDensidadPoblacion() {
+
+	public float calcularDensidadPoblacion() {
 		return poblacion / superficie;
 	}
 }
